requirements.txt
torch==2.2.2
torchvision==0.17.2

opencv-python==4.9.0.80
albumentations==1.4.4
scikit-image==0.22.0

numpy==1.26.4
scipy==1.12.0
pandas==2.2.1

tqdm==4.66.2
tensorboard==2.16.2
rich==13.7.1

pycocotools==2.0.7          
shapely==2.0.3             

PyYAML==6.0.1
hydra-core==1.3.2          

onnx==1.15.0
onnxruntime==1.17.1

matplotlib==3.8.3
seaborn==0.13.2            

black==24.2.0
ruff==0.2.2
pytest==8.0.2